feat: reorganise locale files
