+++ src/app/not-found.tsx
@@ -2,7 +2,7 @@
 
 import { redirect, usePathname } from 'next/navigation';
 
-import { defaultLocale } from '@/lib/config';
+import { defaultLocale } from '@/lib/locale/config';
 
 export default function NotFound() {
   const pathname = usePathname();
