+++ src/middleware.ts
@@ -1,6 +1,11 @@
 import createMiddleware from 'next-intl/middleware';
 
-import { defaultLocale, localePrefix, locales, pathnames } from '@/lib/config';
+import {
+  defaultLocale,
+  localePrefix,
+  locales,
+  pathnames,
+} from '@/lib/locale/config';
 
 export default createMiddleware({
   defaultLocale,
